requirements.txt
async-generator==1.10
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
cryptography==36.0.1
h11==0.13.0
idna==3.3
outcome==1.1.0
Pillow==9.0.1
pycparser==2.21
pyOpenSSL==22.0.0
pytesseract==0.3.8
selenium==4.1.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.8
wincertstore==0.2
wsproto==1.0.0
